fix a bug of building read/written variable just when exiting a method. In my case, a method invocation m(output) reads a variable "output", and I need to know where the output variable is defined.
In my original implementation, I find the earliest previous node
defining variable "output". It is fine in most cases, however, for a
step of exiting a method, the step defining variable "output" could be
located in the method "m" by heap ID. However, the case should be the
earliest previous step outside of method "m".