fix a bug, note that, when we form a path, the end trace node is the suspicious node, however, the start node may not always be the current node!
Consider the case when two read variables are selected in step k, let's
note the step defining the first read variable as step_(k-i), and the
step defining the second as step_(k-j). When jump to step_(k-i) and the
user find it correct, thus, the next suspicious step is step_(k-j). In
this moment, the *current node* is step_(k-i). If we consider form a
path as <step_(k-i), step_(k-j)>, that is wrong. In this case, we should
recoard the path as <step_k, step_(k-j)>