add one more state as INSPECT_DETAIL.
The rationale is as follows: I find it is possible that, given a
dominance relation, the start node has complete correct read/written
vars while the end node has incorrect read var. The case is possible
that some assignment to the read var in the end node is missing. It
could happen either by having the wrong variable in some conditional
step or by the fact that the user just miss writing the code.