fix an important bug
The implementation of l2t record the input variable in the first line in
the method. However, it is usually (at least in Math project) that the
first line in the method cannot be set a breakpoint, e.g., int a;
(statement of variable declaration with definition). Thus, it makes JVM
halt. In my fix, if such case happen, I will record the input variable
in the following breakpoint.